+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+### Added
+
+- Support the `disabled` prop on the `Checkmark` component.
+
 ## [16.3.0] - 2020-01-07
 
 ### Added